Add files via upload
The objective of this project is to present several applications of statistical process control (SPC) in maintaining and improving the quality in coffee shops. It is found that the critical quality characteristic can be observed in the weight of the coffee in every shot of espresso. The goal here is to minimise the variation by constructing a full SPC plan, determining the cause of variation, designing an active out-of-control action plan, and executing the X bar and S chart with the phase I analysis.